+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -3,6 +3,14 @@
 ## 12.27.0b1 (Unreleased)
 
 ### Features Added
+- Added `list_ranges` and `list_ranges_diff` APIs to `ShareFileClient` which return an auto-paging iterable of
+`FileRange` and support continuation tokens and `results_per_page` for enumerating file ranges across multiple
+service responses.
+- Added a new `FileRange` model with `start`, `end`, and `cleared` properties.
+
+### Other Changes
+- Deprecated `ShareFileClient`'s `get_ranges` and `get_ranges_diff` APIs in favor of `list_ranges` and
+`list_ranges_diff`.
 
 ## 12.26.0 (2026-06-08)